refactor: Use flat route convention to colocate code
* refactor: Simplfy conditional login theme switch
