+++ app/routes/_public-md+/_layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from '@remix-run/react'
 
-export default function LandingLayout() {
+export default function PublicLayout() {
   return (
     <article className="container prose py-6 dark:prose-invert">
       <Outlet />
