+++ vite.config.ts
@@ -1,5 +1,6 @@
 import mdx from '@mdx-js/rollup'
 import { vitePlugin as remix } from '@remix-run/dev'
+import { flatRoutes } from 'remix-flat-routes'
 import { defineConfig } from 'vite'
 import tsconfigPaths from 'vite-tsconfig-paths'
 
@@ -8,6 +9,17 @@ export default defineConfig({
     tsconfigPaths(),
     mdx(),
     remix({
+      ignoredRouteFiles: ['**/*'],
+      routes: async (defineRoutes) => {
+        return flatRoutes('routes', defineRoutes, {
+          ignoredRouteFiles: [
+            '.*',
+            '**/*.css',
+            '**/*.test.{js,jsx,ts,tsx}',
+            '**/__*.*',
+          ],
+        })
+      },
       future: {
         v3_fetcherPersist: true,
         v3_relativeSplatPath: true,
